Time: 0 ms (100.00%) | Memory: 8.3 MB (76.99%) - LeetSync

+++ 29-divide-two-integers/divide-two-integers.cpp
@@ -10,7 +10,7 @@ public:
         unsigned int d = abs(divisor);
         unsigned int ans = 0;
         while (n >= d) {
-            short count = 0;
+            int count = 0;
             while (n > (d << (count + 1)))
                 count++;
             ans += (1 << count);
